% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{build-output-class}
\alias{build-output-class}
\title{Output of the pdphmc build process}
\description{
Contains path to binary and also specification of the sampler used and other auxiliary information
}
\section{Slots}{

\describe{
\item{\code{model.name}}{model name}

\item{\code{model.cppcode}}{c++ code}

\item{\code{work.folder}}{folder containing binaries etc}

\item{\code{build.time}}{build.time}

\item{\code{build.ID}}{identifier}

\item{\code{file.name.base}}{absolute path base}

\item{\code{data.vars}}{names and info on DATA variables in model}

\item{\code{par.vars}}{names and info on PARAMETER variables in model}

\item{\code{gen.vars}}{names and info on GENERATED variables in model}

\item{\code{massMatrix}}{type of mass Matrix}

\item{\code{lambda}}{type of event rate}

\item{\code{compile.exit.flag}}{compile exit flag}

\item{\code{binary.path}}{path to the binary}
}}

