% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{remove files associated with pdphmc model}
\usage{
clean(model)
}
\arguments{
\item{model}{S4 class of type `build-output` typically generated by function `build`}
}
\description{
remove files associated with model generated by `build`
}
\examples{
file.copy(system.file('examples/test_case.cpp', package = "pdphmc"),".")
mdl <- pdphmc::build("test_case.cpp")
fit <- pdphmc::run(mdl)
pdphmc::clean(mdl) # will no longer work as binary is deleted...
file.remove("test_case.cpp") #clean up
}
