% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{pdphmc-output-class}
\alias{pdphmc-output-class}
\title{Class containing pdphmc output and diagnostics information}
\description{
pdphmc simulation output class
}
\section{Slots}{

\describe{
\item{\code{modelname}}{model name}

\item{\code{model.obj}}{copy of the build-output object}

\item{\code{niter}}{not currently in use}

\item{\code{chains}}{number of trajectories}

\item{\code{warmup}}{number of samples in warmup mode}

\item{\code{samples}}{number of samples}

\item{\code{pointSamples}}{array containing (point) samples (sample,chain,parameter)}

\item{\code{intSamples}}{array containing integrated samples (sample,chain,parameter)}

\item{\code{eventSamples}}{not currently in use}

\item{\code{diagonstics}}{diagnostics output for each integration leg}

\item{\code{monitor}}{not in use}

\item{\code{intMonitor}}{not in use}

\item{\code{par.vec.names}}{names of parameter vector}

\item{\code{CPUtime}}{CPUtime}

\item{\code{fixedMi}}{inverse mass diagonal fixed}

\item{\code{lastMiDiag}}{last inverse mass diagonal}
}}

